News Flash LAS 2022: YouTube links to edited videos
Right after the conference we had only two videos for the recordings,
one for the first and one for the second day, each 9 hours long. Our
links simply had a time pointer to the beginning of the session
referred to. This is not perfectly reliable.

Now there are also separate videos for each individual session. This
commit switches the YouTube links to these videos.